requirements.txt
certifi==2018.4.16
chardet==3.0.4
future==0.16.0
idna==2.7
peewee==3.5.2
psycopg2-binary==2.7.5
putio.py==8.2.0
python-telegram-bot==10.1.0
requests==2.19.1
six==1.11.0
tus.py==1.3.1
urllib3==1.23
